ci/wheel/requirements.runner.txt
annotated-types==0.6.0
async-timeout==4.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
exceptiongroup==1.2.0
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-monitoring==2.18.0
google-cloud-run==0.10.1
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-core==2.14.6
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0